+++ activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb
@@ -106,6 +106,7 @@ module ActiveRecord
         binary:       { name: "blob" },
         boolean:      { name: "boolean" },
         json:         { name: "json" },
+        jsonb:        { name: "json" },
       }
 
       DEFAULT_PRAGMAS = {